\name{POP.DEF}
\alias{POP.DEF}
\title{population definition table}
\description{
The pop.def table is used to pre-gate and guide the clustering analysis in the classify step.  This one is a hard coded dataframe that is used by default in the analysis if one is not specified by the user.
}
\examples{
POP.DEF
}